Termiando ventana de estrella
- Terminada cajas de estrellas
