+++ mobile_app/lib/src/pages/star_page.dart
@@ -1,9 +1,12 @@
 // ignore_for_file: prefer_const_constructors
 
+import 'dart:ffi';
 import 'dart:math';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:mobile_app/src/pages/pages.dart';
+import 'package:mobile_app/src/providers/menu_providers.dart';
+
 
 class StarPage extends StatefulWidget { //cambia su estado
 
@@ -12,11 +15,6 @@ class StarPage extends StatefulWidget { //cambia su estado
 }
 
 class _AnimatedContainerPageState extends State<StarPage> { 
-  //medidas default
-  double _width = 100.0; 
-  double _height = 100.0;
-  Color _color = Colors.cyan;
-  BorderRadiusGeometry _borderRadius = BorderRadius.circular(5.0);
 
   @override
   Widget build(BuildContext context) { //contenido
@@ -29,111 +27,38 @@ class _AnimatedContainerPageState extends State<StarPage> {
         children: [
             _texto(),
             Divider(),
-            _container()
+            _container(),
           ],
-          ),
-        floatingActionButton: FloatingActionButton( //crear el boton q lo cambia
-          child: Icon(Icons.refresh), //icono de play arrow
-          onPressed: () => null, //si se presiona acciona el void
-          ),
-    );
-
-  
+          )  
+           
+          );
   }
 
   Widget _texto(){
     return Text(
-    'Pulsa las cajas para encontrar la estrella escondida',
+    'Pulsa las cajas para encontrar las estrella escondidas',
     style: TextStyle(
       fontSize: 20,
       color: Colors.black,
     ));
   }
 
-  Widget _contenedor(){
-
-  return AnimatedContainer( //contenedor de animacion
-          duration: Duration(milliseconds: 1000), //la duracion de la animacion
-          curve: Curves.fastOutSlowIn, //cambio de curvas rapido
-          width: _width, //ancho
-          height: _height, //alto
-          decoration: BoxDecoration( //decoracion interna
-            borderRadius: _borderRadius, //borde del radio
-            color: _color //color
-          ),
-        );
-}
 
   Widget _container() {
-    EdgeInsets margen =  EdgeInsets.only(top: 10, left: 15, right: 15, bottom: 20);
     return Container(
       child: Wrap(
         direction: Axis.horizontal,
         children: <Widget>[
-          Container(
-            margin: margen,
-            color: Colors.blue,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("1",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.red,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("2",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.teal,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("3",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.indigo,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("4",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.orange,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("5",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.lightGreen,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("6",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.yellowAccent,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("7",textScaleFactor: 2.5,))
-          ),
-          Container(
-            margin: margen,
-            color: Colors.purple,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("8",textScaleFactor: 2.5,))
-          ),
-          WidgetAnimado()
-          /*Container(
-            margin: margen,
-            color: Colors.amber,
-            width: 100,
-            height: 100,
-            child:Center(child: Text("9",textScaleFactor: 2.5,))
-          ),*/
+          WidgetAnimado("1" , Colors.blue),
+          WidgetAnimado("2" , Colors.red),
+          WidgetAnimado("3" , Colors.teal),
+          WidgetAnimado("4" , Colors.indigo),
+          WidgetAnimado("5" , Colors.orange),
+          WidgetAnimado("6" , Colors.lightGreen),
+          WidgetAnimado("7" , Colors.blueGrey),
+          WidgetAnimado("8" , Colors.purple),
+          WidgetAnimado("9" , Colors.amber),
+         
         ],
       ),
     );
@@ -141,20 +66,33 @@ class _AnimatedContainerPageState extends State<StarPage> {
 
 }
 
-class WidgetAnimado extends StatefulWidget {
-  const WidgetAnimado({Key? key}) : super(key: key);
+class WidgetAnimado extends StatefulWidget {  
+   String numero = " ";
+   MaterialColor colorful = Colors.deepOrange;
+   WidgetAnimado(String s, MaterialColor color) {
+     this.numero = s;
+     this.colorful = color;
+   }
 
   @override
-  State<WidgetAnimado> createState() => _ContainerState();
+  State<WidgetAnimado> createState() => _ContainerState(Container, numero, colorful);
 }
 
 class _ContainerState extends State<WidgetAnimado> {
   bool selected = false;
+  String numero = "0";
+  MaterialColor colorful = Colors.deepOrange;
+  
+  _ContainerState(Type container, String number, MaterialColor colorido){
+    this.numero = number;
+    this.colorful = colorido;
+  }
 
   @override
   Widget build(BuildContext context) {
-    return AnimatedScale(
-            scale: selected ? 1.10 : 1.00,
+    return 
+    AnimatedScale(
+            scale: selected ? 1.20 : 1.00,
             duration: const Duration(seconds: 1),
             curve: Curves.fastOutSlowIn,
             child: GestureDetector(
@@ -163,17 +101,39 @@ class _ContainerState extends State<WidgetAnimado> {
                   selected = !selected;
                 });
               },
-              child: Container(
-            margin: EdgeInsets.only(top: 10, left: 15, right: 15, bottom: 20),
-            color: Colors.amber,
+              child:  _posiciones(selected)
+              ),
+            );
+  }
+
+
+  Widget _listaContainer(BuildContext context) {
+    return Container(
+            margin: EdgeInsets.only(top: 30, left: 15, right: 15, bottom: 20),
+            color: colorful,
             width: 100,
             height: 100,
-            child:Center(child: Text("9",textScaleFactor: 2.5,))
-              ),
-            ),
+            child:Center(child: Text(numero,textScaleFactor: 2.5,))
+    );
+      
+    }
+
+    Widget _posiciones(bool selected){
+    return Stack(
+      children: <Widget>[    
+          AnimatedContainer(
+            duration: Duration(milliseconds: 1000),
+            curve: Curves.decelerate,
+            margin: selected ? EdgeInsets.only(top: 125, left: 25) : EdgeInsets.only(top: 20, left: 25),
+            child: Icon(
+              Icons.star,
+              color: Colors.yellowAccent,
+              size: 75,
+              )
+          ),
+          _listaContainer(context),
+        ],
     );
   }
-}
-
-
+  }
