+++ mobile_app/lib/src/routes/routes.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/src/pages/pages.dart';
-import 'package:mobile_app/src/pages/staro_page.dart';
 
 Map<String, WidgetBuilder> getRoutes() {
   return <String, WidgetBuilder>{
     '/': (BuildContext context) => HomePage(),
     'gallery': (BuildContext context) => GalleryPage(),
-    'star': (BuildContext context) => StaroPage(),
+    'star': (BuildContext context) => StarPage(),
     'pong': (BuildContext context) => PongGame(),
     'language': (BuildContext context) => LocalizationSystemPage()
   };
